requirements.txt
appdirs==1.4.4
asgiref==3.2.10
black==20.8b1
cachetools==4.2.1
certifi==2020.12.5
cffi==1.14.5
chardet==4.0.0
click==7.1.2
colorama==0.4.4
coverage==5.3
cryptography==3.4.7
defusedxml==0.7.1
Django==3.1
django-allauth==0.43.0
django-classy-tags==2.0.0
django-cookie-law==2.0.5
django-debug-toolbar==3.2
django-environ==0.4.5
django-extensions==3.1.0
django-recaptcha==2.0.6
django-rosetta==0.9.5
django-storages==1.11.1
django-taggit==1.3.0
django-widget-tweaks==1.4.8
factory-boy==3.1.0
Faker==8.1.0
google-api-core==1.26.3
google-auth==1.29.0
google-cloud-core==1.6.0
google-cloud-storage==1.37.1
google-crc32c==1.1.2
google-resumable-media==1.2.0
googleapis-common-protos==1.53.0
gunicorn==20.1.0
idna==2.10
isort==5.7.0
mypy-extensions==0.4.3
oauthlib==3.1.0
packaging==20.9
pathspec==0.8.1
Pillow==8.1.2
polib==1.1.1
protobuf==3.15.8
psycopg2-binary==2.8.6
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.20
PyJWT==2.0.1
pyparsing==2.4.7
pyprof2calltree==1.4.5
pytest-runner==5.3.0
python-dateutil==2.8.1
python3-openid==3.2.0
pytz==2021.1
regex==2021.4.4
requests==2.25.1
requests-oauthlib==1.3.0
rsa==4.7.2
selenium==3.141.0
six==1.15.0
sqlparse==0.4.1
tblib==1.7.0
text-unidecode==1.3
toml==0.10.2
typed-ast==1.4.3
typing-extensions==3.7.4.3
urllib3==1.26.4
